+++ checked.go
@@ -153,5 +153,6 @@ var (
 		"zaplog":                 true,
 		"consul":                 true,
 		"prom_remote_write":      true,
+		"pythond":                true,
 	}
 )