+++ http/api_dca.go
@@ -187,6 +187,13 @@ type saveConfigParam struct {
 
 // auth middleware.
 func dcaAuthMiddleware(c *gin.Context) {
+	fullPath := c.FullPath()
+	for _, uri := range ignoreAuthURI {
+		if uri == fullPath {
+			c.Next()
+			return
+		}
+	}
 	tokens := c.Request.Header["X-Token"]
 	context := &dcaContext{c: c}
 	if len(tokens) == 0 {
@@ -562,4 +569,107 @@ func dcaTestPipelines(c *gin.Context) {
 	}
 
 	context.success(parsedText)
+}
+
+type rumQueryParam struct {
+	ApplicatinID string `form:"app_id"`
+	Env          string `form:"env"`
+	Version      string `form:"version"`
+}
+
+// upload sourcemap
+// curl -X POST 'http://localhost:9531/v1/rum/sourcemap?app_id=app_xxxx&env=release&version=1.0.1'
+// 			-F "file=@/tmp/code.zip"
+// 			-H "Content-Type: multipart/form-data".
+func dcaUploadSourcemap(c *gin.Context) {
+	context := getContext(c)
+
+	var param rumQueryParam
+
+	if c.ShouldBindQuery(&param) != nil {
+		context.fail(dcaError{ErrorCode: "query.parse.error", ErrorMsg: "query string parse error"})
+		return
+	}
+
+	if (len(param.ApplicatinID) == 0) || (len(param.Env) == 0) || (len(param.Version) == 0) {
+		context.fail(dcaError{ErrorCode: "query.param.required", ErrorMsg: "app_id, env, version required"})
+		return
+	}
+
+	file, err := c.FormFile("file")
+	if err != nil {
+		l.Errorf("get file failed: %s", err.Error())
+		context.fail(dcaError{ErrorCode: "upload.file.required", ErrorMsg: "make sure the file was uploaded correctly"})
+		return
+	}
+
+	fileName := GetSourcemapZipFileName(param.ApplicatinID, param.Env, param.Version)
+	rumDir := GetRumSourcemapDir()
+	if !path.IsDir(rumDir) {
+		if err := os.MkdirAll(rumDir, datakit.ConfPerm); err != nil {
+			context.fail(dcaError{
+				ErrorCode: "dir.create.failed",
+				ErrorMsg:  "rum dir created failed",
+			})
+			return
+		}
+	}
+	dst := filepath.Clean(filepath.Join(rumDir, fileName))
+
+	// check filename
+	if !strings.HasPrefix(dst, rumDir) {
+		context.fail(dcaError{
+			ErrorCode: "param.invalid",
+			ErrorMsg:  "invalid param, should not contain illegal char, such as  '../, /'",
+		})
+		return
+	}
+
+	if err := c.SaveUploadedFile(file, dst); err != nil {
+		l.Errorf("save upload file error: %s", err.Error())
+		context.fail(dcaError{ErrorCode: "upload.file.error", ErrorMsg: "upload failed"})
+		return
+	}
+	updateSourcemapCache(dst)
+	context.success(fmt.Sprintf("uploaded to %s!", fileName))
+}
+
+func dcaDeleteSourcemap(c *gin.Context) {
+	context := getContext(c)
+
+	var param rumQueryParam
+
+	if c.ShouldBindQuery(&param) != nil {
+		context.fail(dcaError{ErrorCode: "query.parse.error", ErrorMsg: "query string parse error"})
+		return
+	}
+
+	if (len(param.ApplicatinID) == 0) || (len(param.Env) == 0) || (len(param.Version) == 0) {
+		context.fail(dcaError{ErrorCode: "query.param.required", ErrorMsg: "app_id, env, version required"})
+		return
+	}
+
+	fileName := GetSourcemapZipFileName(param.ApplicatinID, param.Env, param.Version)
+	rumDir := GetRumSourcemapDir()
+	zipFilePath := filepath.Clean(filepath.Join(rumDir, fileName))
+
+	// check filename
+	if !strings.HasPrefix(zipFilePath, rumDir) {
+		context.fail(dcaError{
+			ErrorCode: "param.invalid",
+			ErrorMsg:  "invalid param, should not contain illegal char, such as  '../, /'",
+		})
+		return
+	}
+
+	if err := os.Remove(zipFilePath); err != nil {
+		l.Errorf("delete zip file failed: %s, %s", zipFilePath, err.Error())
+		context.fail(dcaError{
+			ErrorCode: "delete.error",
+			ErrorMsg:  "delete sourcemap file failed.",
+		})
+		return
+	}
+	deleteSourcemapCache(zipFilePath)
+	context.success("delete file successfully")
 }