+++ plugins/inputs/ddtrace/input.go
@@ -8,8 +8,8 @@ import (
 	"gitlab.jiagouyun.com/cloudcare-tools/cliutils/logger"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/http"
+	"gitlab.jiagouyun.com/cloudcare-tools/datakit/internal/trace"
 	"gitlab.jiagouyun.com/cloudcare-tools/datakit/plugins/inputs"
-	"gitlab.jiagouyun.com/cloudcare-tools/datakit/plugins/inputs/trace"
 )
 
 var (
@@ -39,7 +39,8 @@ var (
 `
 	customerKeys []string
 	ddTags       map[string]string
-	log          = logger.DefaultSLogger(inputName)
+	log                         = logger.DefaultSLogger(inputName)
+	_            inputs.InputV2 = &Input{}
 )
 
 var (
@@ -69,11 +70,11 @@ func (*Input) SampleConfig() string {
 	return ddtraceSampleConfig
 }
 
-func (i *Input) AvailableArchs() []string {
+func (*Input) AvailableArchs() []string {
 	return datakit.AllArch
 }
 
-func (i *Input) SampleMeasurement() []inputs.Measurement {
+func (*Input) SampleMeasurement() []inputs.Measurement {
 	return []inputs.Measurement{&DDTraceMeasurement{}}
 }